fix(indexer): SMI-5286 1c governance — fix facet retirement (C-1) + harden bisection
Resolves all governance-review findings on 8ddf7d34:

- C-1 (Critical): bisectCurrentFacet now RETIRES a saturated top-level facet
  (facetIndex++) before pushing its halves. Previously a top-level facet that
  stayed saturated was re-queried after its sub-ranges drained → infinite
  re-crawl with facets_completed stuck at 0. Two regression tests added
  (unit retirement + integration persistent-saturation, budget-bounded).
- Open-ended bisection ceiling (4 MiB): an always-saturating open-ended facet
  doubled forever (never reaching lo===hi); bisectFacet now returns null past
  the ceiling so the tail terminates as truncated. Test added.
- M-1: a page error on a range is now counted in truncated_repo_count + logged
  (was silently skipped) and the crawl advances. Test added.
- M-2: BackfillSummary gains current_facet + pending_subrange_count; the
  workflow terminal condition now keys on current_facet=='done' (authoritative)
  instead of facets_remaining==0, which reads 0 while the last facet's
  bisected sub-ranges still drain.
- L-2: split the facet state-machine tests into
  backfill-checkpoint.statemachine.test.ts (both files now <500 lines).

Indexer suite 471 green; lint/typecheck/format/audit:standards clean.
